Refactor back to plugin.py file
