+++ ckanext/pages/i18n/de/LC_MESSAGES/ckanext-pages.mo
