+++ ckanext/pages/plugin.py
@@ -4,17 +4,17 @@ from html import escape as html_escape
 
 from six.moves.urllib.parse import quote
 
-import ckantoolkit as tk
+from ckan.plugins import toolkit as tk
 
 import ckan.plugins as p
 from ckan.lib.helpers import build_nav_main as core_build_nav_main
 
 from ckanext.pages import actions, db
 from ckanext.pages import auth
+from ckanext.pages import blueprint
+from ckanext.pages import cli
 
 from ckan.lib.plugins import DefaultTranslation
-
-from ckanext.pages.plugin.flask_plugin import MixinPlugin
 
 
 log = logging.getLogger(__name__)
@@ -90,28 +90,36 @@ class PagesPluginBase(p.SingletonPlugin, DefaultTranslation):
     p.implements(p.ITranslation, inherit=True)
 
 
-class PagesPlugin(PagesPluginBase, MixinPlugin):
+class PagesPlugin(PagesPluginBase):
     p.implements(p.IConfigurer, inherit=True)
     p.implements(p.ITemplateHelpers, inherit=True)
     p.implements(p.IActions, inherit=True)
     p.implements(p.IAuthFunctions, inherit=True)
     p.implements(p.IConfigurable, inherit=True)
+    p.implements(p.IBlueprint)
+    p.implements(p.IClick)
+
+    def get_blueprint(self):
+        return [blueprint.pages]
+
+    def get_commands(self):
+        return cli.get_commands()
 
     def update_config(self, config):
         self.organization_pages = tk.asbool(config.get('ckanext.pages.organization', False))
         self.group_pages = tk.asbool(config.get('ckanext.pages.group', False))
 
-        tk.add_template_directory(config, '../theme/templates_main')
+        tk.add_template_directory(config, 'theme/templates_main')
         if self.group_pages:
-            tk.add_template_directory(config, '../theme/templates_group')
+            tk.add_template_directory(config, 'theme/templates_group')
         if self.organization_pages:
-            tk.add_template_directory(config, '../theme/templates_organization')
+            tk.add_template_directory(config, 'theme/templates_organization')
 
-        tk.add_resource('../assets', 'pages')
+        tk.add_resource('assets', 'pages')
 
-        tk.add_public_directory(config, '../assets/')
-        tk.add_public_directory(config, '../assets/vendor/ckeditor/')
-        tk.add_public_directory(config, '../assets/vendor/ckeditor/skins/moono-lisa')
+        tk.add_public_directory(config, 'assets/')
+        tk.add_public_directory(config, 'assets/vendor/ckeditor/')
+        tk.add_public_directory(config, 'assets/vendor/ckeditor/skins/moono-lisa')
 
     def get_helpers(self):
         return {
